Move waitFor utiity into src/utils.js so can be used elsewhere.
